requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
autograd==1.8.0
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.4
blis==1.2.1
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.1.8
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
durationpy==0.10
emoji==2.14.1
environs==11.2.1
executing==2.2.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fnllm==0.3.0
fonttools==4.58.0
frozenlist==1.6.2
fsspec==2025.5.1
future==1.0.0
gensim==4.3.3
google-auth==2.40.2
googleapis-common-protos==1.70.0
graphrag==2.3.0
graspologic==3.4.1
graspologic-native==1.2.5
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
hyppo==0.4.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
isodate==0.7.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jq==1.8.0
json-repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
lancedb==0.17.0
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.44
language-data==1.3.0
llvmlite==0.44.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
murmurhash==1.0.13
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.84.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
posthog==4.2.0
pot==0.9.5
preshed==3.0.10
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
pylance==0.20.0
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.6.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.41
srsly==2.5.1
starlette==0.45.3
statsmodels==0.14.4
sympy==1.14.0
tenacity==9.1.2
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.4
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
umap-learn==0.5.7
unstructured==0.17.2
unstructured-client==0.36.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
